fix multi thread start <tshark> error --> 多个线程同时启动tshark会报错，所以这里的策略是启动一个tshark时暂停2s，然后启动下一个tshark，之后需要优化一下；延时解析失败...

+++ src/main/java/com/zjucsc/application/config/Common.java
@@ -39,7 +39,7 @@ public class Common {
     /**
      * 采集器的时延，key是采集器的ID，value是解析出来的最大时延
      */
-    public static ConcurrentHashMap<Integer,Long> COLLECTOR_DELAY_MAP = new ConcurrentHashMap<>();
+    public static ConcurrentHashMap<Integer,Integer> COLLECTOR_DELAY_MAP = new ConcurrentHashMap<>();
 
     /**
      * 异常数【五元组和过滤规则】
